prefer int to size_t

+++ src/terrain/terrain.cpp
@@ -2,13 +2,13 @@
 
 #include <iostream>
 
-Terrain::Terrain(size_t x, size_t y)
+Terrain::Terrain(unsigned int x, unsigned int y)
     : size_x(x), size_y(y)
 {
-    heightmap = std::vector<std::vector<size_t>>(size_x);
+    heightmap = std::vector<std::vector<unsigned int>>(size_x);
     for(int i=0; i<size_x; i++)
     {
-        heightmap[i] = std::vector<size_t>(size_y);
+        heightmap[i] = std::vector<unsigned int>(size_y);
     }
 
     std::cout << heightmap.capacity() << std::endl;
@@ -17,7 +17,7 @@ Terrain::Terrain(size_t x, size_t y)
 Terrain::Terrain(Heightmap _heightmap):
 size_x(_heightmap.capacity()), size_y(_heightmap.at(0).capacity())
 {
-    heightmap = std::vector<std::vector<size_t>>(size_x);
+    heightmap = std::vector<std::vector<unsigned int>>(size_x);
     for(int i=0; i<size_x; i++)
     {
         heightmap[i] = _heightmap.at(i);
